amazon-python-scrapy-scraper/requirements.txt
amqp==5.1.1
async-timeout==4.0.3
attrs==23.1.0
Automat==22.10.0
billiard==4.1.0
blinker==1.6.2
boto3==1.26.157
botocore==1.29.157
celery==5.3.1
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
constantly==15.1.0
contourpy==1.1.0
crochet==2.0.0
cryptography==41.0.1
cssselect==1.2.0
cycler==0.11.0
distlib==0.3.6
fake-useragent==1.1.3
Faker==19.2.0
filelock==3.12.0
Flask==2.3.2
flask-mysql-connector==1.1.0
fonttools==4.40.0
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
itemadapter==0.8.0
itemloaders==1.1.0
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.2.0
json5==0.9.14
kiwisolver==1.4.4
kombu==5.3.1
lxml==4.9.2
MarkupSafe==2.1.3
matplotlib==3.7.1
mysql-connector-python==8.0.33
nltk==3.8.1
numpy==1.25.0
packaging==23.1
pandas==2.0.2
parsel==1.8.1
Pillow==9.5.0
platformdirs==3.5.1
plotly==5.15.0
prompt-toolkit==3.0.39
Protego==0.2.1
protobuf==3.20.3
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
PyDispatcher==2.0.7
pyOpenSSL==23.2.0
pyparsing==3.1.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
queuelib==1.6.2
redis==4.6.0
regex==2023.6.3
requests==2.31.0
requests-file==1.5.1
rq==1.15.1
s3transfer==0.6.1
scikit-learn==1.2.2
scipy==1.10.1
scrapeops-scrapy==0.5.2
scrapeops-scrapy-proxy-sdk==1.0
Scrapy==2.9.0
scrapy-fake-useragent==1.4.4
service-identity==21.1.0
six==1.16.0
tenacity==8.2.2
threadpoolctl==3.1.0
tld==0.13
tldextract==3.4.4
tqdm==4.65.0
Twisted==22.10.0
typing_extensions==4.6.3
tzdata==2023.3
urllib3==1.26.16
vine==5.0.0
virtualenv==20.23.0
w3lib==2.1.1
wcwidth==0.2.6
Werkzeug==2.3.6
wordcloud==1.9.2
wrapt==1.15.0
zope.interface==6.0
